refactor: change u64 to i64

+++ src/app/analysis/git_analysis/file_analysis.rs
@@ -2,7 +2,7 @@ use crate::infrastructure::git::git_file_history;
 use core_model::url_format;
 use git_scanner::flare::FlareTreeNode;
 
-pub fn analysis(url: &str, git_years: u64) -> FlareTreeNode {
+pub fn analysis(url: &str, git_years: f64) -> FlareTreeNode {
     let local_path = url_format::uri_to_path(url);
     let tree_node = git_file_history::by_path(local_path, git_years);
